% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_prob.R
\name{change_current_prob_to_cycle_prob}
\alias{change_current_prob_to_cycle_prob}
\title{Title
This function converts the given probabilities to match the model's cycle length}
\usage{
change_current_prob_to_cycle_prob(prob, year_prob, cycle_length)
}
\arguments{
\item{prob}{probability}

\item{year_prob}{current terms of the given probability in years}

\item{cycle_length}{cycle length in years}
}
\value{
cycle_prob
}
\description{
Title
This function converts the given probabilities to match the model's cycle length
}
\examples{
drug_arm_5_years<-matrix(data = c(0,0.029,0.087*0.59,0,0,0.1,0,0.261*0.59,0,0,0,0,0,
0.325,0,0,0,0,1,0,0,0,0,0,1),nrow = 5,ncol = 5,byrow = TRUE)
drug_arm_5_years<-change_current_prob_to_cycle_prob(drug_arm_5_years,1,(1/12))
}
