% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_both_arms_sim.R
\name{check_variables_compatibility}
\alias{check_variables_compatibility}
\title{Title
This function ensures the model's input are compatible}
\usage{
check_variables_compatibility(trans_prob_mat, util, cost, inital_state_vec,
  dead_state_num)
}
\arguments{
\item{trans_prob_mat}{transition matrix}

\item{util}{utility vector}

\item{cost}{cost vector}

\item{inital_state_vec}{initial stste vector}

\item{dead_state_num}{number of dead states}
}
\value{
warning message
}
\description{
Title
This function ensures the model's input are compatible
}
