% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_main_func.R
\name{cond_prob_calc}
\alias{cond_prob_calc}
\title{Title
conditional probability sd calculation}
\usage{
cond_prob_calc(prob_matrix, discount_rate, cost, util, currentT,
  trans_prob_mat, dead_state_num, cycle_length, ly_val_mat)
}
\arguments{
\item{prob_matrix}{state probabilities for each time}

\item{discount_rate}{discount rate in percentages}

\item{cost}{cost vector}

\item{util}{utility vector}

\item{currentT}{current time}

\item{trans_prob_mat}{transition matrix}

\item{dead_state_num}{number of dead states}

\item{cycle_length}{cycle length in years}

\item{ly_val_mat}{LYs matrix}
}
\value{
list
}
\description{
Title
conditional probability sd calculation
}
