% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_main_func.R
\name{cond_prob_calc_1}
\alias{cond_prob_calc_1}
\title{Title
part one conditional probability sd calculation}
\usage{
cond_prob_calc_1(variable, discount_rate, prob_matrix, currentT, t, i,
  cond_prob_temp)
}
\arguments{
\item{variable}{LYs/QALYs/Costs}

\item{discount_rate}{discount rate in percentages}

\item{prob_matrix}{state probabilities for each time}

\item{currentT}{current time}

\item{t}{time t iterator}

\item{i}{state i iterator}

\item{cond_prob_temp}{temp conditional probability}
}
\value{
list
}
\description{
Title
part one conditional probability sd calculation
}
