% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_discount_rate.R
\name{covert_discount_rate}
\alias{covert_discount_rate}
\title{Title
This function converts an annual discount rate to the proper discount rate
based on the cycle length of the model}
\usage{
covert_discount_rate(annual_discount_rate, cycle_length)
}
\arguments{
\item{annual_discount_rate}{annual discount rate in percentages}

\item{cycle_length}{cycle length in years}
}
\value{
new_discount_rate
}
\description{
Title
This function converts an annual discount rate to the proper discount rate
based on the cycle length of the model
}
