% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_prob.R
\name{extract_by_age}
\alias{extract_by_age}
\title{Title
This function extract the all-cause mortality probabilities
using the inital age and time-horizon}
\usage{
extract_by_age(initial_age, time_horizon, data_table)
}
\arguments{
\item{initial_age}{initial age}

\item{time_horizon}{null for lifetime horizon or number of years}

\item{data_table}{data frame containing probabilities}
}
\value{
needed_data
}
\description{
Title
This function extract the all-cause mortality probabilities
using the inital age and time-horizon
}
