% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_main_func.R
\name{f1_sum}
\alias{f1_sum}
\title{Title
part one of the sd calculation}
\usage{
f1_sum(prob_matrix, discount_rate, currentT, variable)
}
\arguments{
\item{prob_matrix}{state probabilities for each time}

\item{discount_rate}{discount rate in percentage}

\item{currentT}{current time}

\item{variable}{LYs/QALYs/Costs}
}
\value{
list
}
\description{
Title
part one of the sd calculation
}
