% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_trans_matrix.R
\name{fill_trans_matrix_val}
\alias{fill_trans_matrix_val}
\title{Title
This function embeds the all-cause mortality probabilities in the transition matrices}
\usage{
fill_trans_matrix_val(trans_matrix, prob_of_dying)
}
\arguments{
\item{trans_matrix}{transition matrix}

\item{prob_of_dying}{all-cause mortality probability}
}
\value{
trans_matrix
}
\description{
Title
This function embeds the all-cause mortality probabilities in the transition matrices
}
