% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_main_func.R
\name{final_res}
\alias{final_res}
\title{Title
sum of conditional probability one and two}
\usage{
final_res(prob_matrix, currentT, cond_prob_1, cond_prob_2)
}
\arguments{
\item{prob_matrix}{state probabilities for each time}

\item{currentT}{current time}

\item{cond_prob_1}{temp conditional probability 1}

\item{cond_prob_2}{temp conditional probability 2}
}
\value{
number
}
\description{
Title
sum of conditional probability one and two
}
