% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_main_func.R
\name{main_sd_function}
\alias{main_sd_function}
\title{Title
part two of the sd calculation}
\usage{
main_sd_function(prob_matrix, cost, util, trans_prob_mat, dead_state_num,
  cycle_length, discount_rate)
}
\arguments{
\item{prob_matrix}{state probabilities for each time}

\item{cost}{cost vector}

\item{util}{utility vector}

\item{trans_prob_mat}{probability matrix}

\item{dead_state_num}{number of dead states}

\item{cycle_length}{cycle length in years}

\item{discount_rate}{discount rate in percentages}
}
\value{
list
}
\description{
Title
part two of the sd calculation
}
